Add files via upload
YouTube, the free video sharing website, has emerged as one of the most versatile platforms for sharing knowledge and information. It has become one of the most primary source of entertainments in the present arena and is being browsed by millions of people every second across the globe. On a larger and more useful scale, it has catered to the limpid expression of thoughts and creative ideas by people in a myriad of fields. Although YouTube serves as an excellent platform for sharing a wide variety of videos, one important factor that determines the impact a video makes is its popularity. Since there are almost uncountable videos shared here every day, the popularity of a video depends on a large number of factors. One such important aspect is the ‘Trending’ section which brings together number of videos from different countries under a single heading. Through this project we have tried to focus on this parameter and what factors contribute towards bringing a video into this very section. During our brainstorming sessions, we tried to analyze the existing algorithms that might be governing the functioning of the ‘Trending’ section. We also tried to come up with some factors which we felt might be vital based on our thought process and the prior knowhow of the subject matter. Based on a schematic list of all possible factors that the team could come up with, we followed a systematic approach for data analysis utilizing all the classroom techniques that we have imbibed.